ember-addon skeleton
