+++ .travis.yml
@@ -1,5 +1,11 @@
+---
 language: node_js
+
 sudo: false
+
+cache:
+  directories:
+    - node_modules
 
 before_install:
   - "npm config set spin false"
@@ -9,9 +15,9 @@ install:
   - npm install -g bower
   - npm install
   - bower install
-  - ./bin/select-ember-version.js
 
-script: npm run build && npm run test
+script:
+  - npm test
 
 env:
   matrix:
@@ -21,6 +27,3 @@ env:
     - EMBER_CHANNEL=canary HTMLBARS=true
   global:
     secure: DBzfaR7q4J+dRPQ/b8xRzfDXeAJUdUMKsiwTYTvrGz9n+N2Oyf91EhU11HXQ5h7V3AX6ATIF4QJ3EXjBdTAa7q7Ss+xQM3Oag/KaMjEdG3PK5pG8KNbXyFKuvKZRKm6R/6sJA924gqLECfOpvw2fHVQyZhOirKft2GLPJL62VDs=
-
-after_success:
-  - "./bin/bower-build"
