+++ addon/lib/system/inflections.js
