+++ addon/lib/system/string.js
