+++ bower.json
@@ -2,20 +2,16 @@
   "name": "ember-inflector",
   "version": "1.4.0",
   "dependencies": {
-    "ember": ">= 1.0.0 < 2.0.0",
-    "handlebars": "2.0.0"
-  },
-  "devDependencies": {
-    "jquery": "~1.11.1",
-    "handlebars": "~2.0.0",
-    "loader.js": "~1.0.0"
-  },
-  "ignore": [
-    "dist/ember-inflector.jshint.js",
-    "dist/tests.js"
-  ],
-  "resolutions": {
-    "ember": "beta",
-    "handlebars": "2.0.0"
+    "jquery": "^1.11.1",
+    "ember": "1.10.0",
+    "ember-data": "1.0.0-beta.15",
+    "ember-resolver": "~0.1.12",
+    "loader.js": "ember-cli/loader.js#3.2.0",
+    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
+    "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.0.2",
+    "ember-qunit": "0.2.8",
+    "ember-qunit-notifications": "0.0.7",
+    "qunit": "~1.17.1"
   }
 }
