+++ config/environment.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(/* environment, appConfig */) {
+  return { };
+};
