--- lib/ember-addon/blueprints/ember-inflector/index.js
@@ -1,13 +1,0 @@
-'use strict';
-
-module.exports = {
-  normalizeEntityName: function() {
-    // this prevents an error when the entityName is
-    // not specified (since that doesn't actually matter
-    // to us
-  },
-
-  afterInstall: function() {
-    return this.addBowerPackageToProject('ember-inflector', '1.3.0');
-  }
-};
