--- packages/ember-inflector/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "ember-inflector",
-  "summary": "Ember Inflections",
-  "descrption": "",
-  "homepage": "http://github.com/emberjs/ember.js",
-  "author": ["Stefan Penner"],
-  "version": "1.0.0-beta.1",
-
-  "directories":{
-    "lib": "lib"
-  },
-
-  "dependencies": {
-    "ember-runtime": "1.0.0"
-  },
-
-  "bpm:build": {
-    "bpm_lib.js": {
-      "files": ["lib"],
-      "modes": "*"
-    }
-  }
-}
